requirements.txt
viam-sdk
random
numpy